+++ FluentFTP/Helpers/FtpListParser.cs
@@ -82,13 +82,13 @@ namespace FluentFTP.Helpers {
 				else
 				{
 					CurrentParser = FtpParser.Unix;
-					client.LogStatus(FtpTraceLevel.Warn, "Cannot auto-detect listing parser for system '" + system + "', using Unix parser");
+					((IInternalFtpClient)client).LogStatus(FtpTraceLevel.Warn, "Cannot auto-detect listing parser for system '" + system + "', using Unix parser");
 				}
 			}
 
 			DetectedParser = CurrentParser;
 
-			client.LogStatus(FtpTraceLevel.Verbose, "Listing parser set to: " + DetectedParser.ToString());
+			((IInternalFtpClient)client).LogStatus(FtpTraceLevel.Verbose, "Listing parser set to: " + DetectedParser.ToString());
 		}
 
 		/// <summary>
@@ -200,7 +200,7 @@ namespace FluentFTP.Helpers {
 
 				// use the initially set parser (from SYST)
 				if (IsParserValid(CurrentParser, files)) {
-					client.LogStatus(FtpTraceLevel.Verbose, "Confirmed format " + CurrentParser.ToString());
+					((IInternalFtpClient)client).LogStatus(FtpTraceLevel.Verbose, "Confirmed format " + CurrentParser.ToString());
 					ParserConfirmed = true;
 					return;
 				}
@@ -208,14 +208,14 @@ namespace FluentFTP.Helpers {
 				foreach (var p in parsers) {
 					if (IsParserValid(p, files)) {
 						CurrentParser = p;
-						client.LogStatus(FtpTraceLevel.Verbose, "Detected format " + CurrentParser.ToString());
+						((IInternalFtpClient)client).LogStatus(FtpTraceLevel.Verbose, "Detected format " + CurrentParser.ToString());
 						ParserConfirmed = true;
 						return;
 					}
 				}
 
 				CurrentParser = FtpParser.Unix;
-				client.LogStatus(FtpTraceLevel.Verbose, "Could not detect format. Using default " + CurrentParser.ToString());
+				((IInternalFtpClient)client).LogStatus(FtpTraceLevel.Verbose, "Could not detect format. Using default " + CurrentParser.ToString());
 			}
 		}